server/requirement.txt
pillow==10.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
APScheduler==3.10.4

fastapi==0.115.0
fastapi-cache2==0.2.2
pydantic-settings==2.5.2
tortoise-orm==0.21.6

aiosmtplib==3.0.2
aiofiles==24.1.0

asyncpg==0.30.0
aiomysql==0.2.0
redis==5.1.1

oss2==2.19.0
qiniu==7.13.2
cos-python-sdk-v5==1.9.31

alibabacloud_dysmsapi20170525==3.0.0
tencentcloud-sdk-python==3.0.1233

python-weixin==0.5.0
wechatpayv3==1.2.52
python-alipay-sdk==3.3.0

requests==2.31.0
beautifulsoup4==4.12.3

pymilvus[milvus-lite] == 2.5.1
torch==2.9.1
torchvision==0.24.1